RED deckShouldContain13CardsOfEachSuit
I must put what was previously in the constructor in a separate method to be able to use the mockito spy. I also have to create a separate createCard method to be able to determine the number of times that method is run and with which arguments to avoid having to use the getters in Card. Perhaps a factory would be an alternative, but it would also increase the difficulty of the testing, so I'll go with this for now